feat(ui-reskin-backend-stubs): frontend DTOs, ProfileViewModel, wiring progress + displayName (p3)
backend/src/models.rs -- fix: progress f32->f64 (FLOAT8 mismatch)
backend/src/handlers/collections.rs -- fix: progress calc f64
frontend/data/api/ApiModels.kt -- CollectionDto(lastStudied, progress), UserDto, LearningCompleteRequest
frontend/data/api/ApiClient.kt -- getMe(), patchLearningComplete()
frontend/data/repository/CollectionRepository.kt -- markStudied()
frontend/data/repository/ProfileRepository.kt (new)
frontend/ProfileViewModel.kt (new)
frontend/di/AppModule.kt -- ProfileRepository, ProfileViewModel; LearningViewModel +collectionRepo
frontend/ProfileScreen.kt -- displayName + email z ProfileViewModel
frontend/CollectionsScreen.kt -- TrackBar(progress = collection.progress)
frontend/FlashcardsScreen.kt -- POSTES stat z collection.progress
frontend/LearningViewModel.kt -- stop() wywoluje markStudied
frontend/CollectionFormScreen.kt -- fix: return@onClick -> ?.let{}

+++ backend/src/models.rs
@@ -28,7 +28,7 @@ pub struct Collection {
     pub target_language: String,
     pub created_at: DateTime<Utc>,
     pub last_studied: Option<DateTime<Utc>>,
-    pub progress: f32,
+    pub progress: f64,
 }
 
 #[derive(Debug, Deserialize)]
